Zipped Folders of Entire Masters Project Added
4 Folders present inside.
ROS-bootstrap: Is the files for starting a local website from visual studio
taigo_pc.zip: Initial files by Sara Cooper for robot control, not used.
Master_project.zip: All the files made for the project's ROS control. As well as starting the gazebo simulation.

The entire Project with all downloaded functionality to do with the tiago is required. Downloading tiago files and Gazebo funtionality can be found on ROS.org